fix: safari horizontal scroll with content overflow APP-1 (#2328)
