+++ web-components/src/components/section/index.tsx
@@ -37,53 +37,60 @@ export interface SectionProps {
 }
 
 interface StyleProps {
-  withSlider: boolean;
+  withSlider?: boolean;
   topRight: boolean;
   titleLineHeight?: string;
   titleColor?: string;
   titleAlign?: 'left' | 'right' | 'inherit' | 'center' | 'justify' | undefined;
 }
 
-const Root = styled(Box, {
+export const Root = styled(Box, {
   name: 'RegenSection',
   shouldForwardProp: prop =>
     prop !== 'withSlider' &&
     prop !== 'visibleOverflow' &&
     prop !== 'isPaddingTopMobile',
 })<{
-  withSlider: boolean;
-  visibleOverflow: boolean;
-  isPaddingTopMobile: boolean;
-}>(({ theme, withSlider, visibleOverflow, isPaddingTopMobile }) => ({
-  maxWidth: theme.breakpoints.values.lg,
-  margin: '0 auto',
-  [theme.breakpoints.up('lg')]: {
-    overflow: visibleOverflow ? 'visible' : 'hidden',
-  },
-  [theme.breakpoints.down('lg')]: {
-    overflow: 'hidden',
-  },
-  [theme.breakpoints.up('sm')]: {
-    paddingTop: theme.spacing(22.25),
-  },
-  [theme.breakpoints.up('md')]: {
-    paddingRight: theme.spacing(37.5),
-    paddingLeft: theme.spacing(37.5),
-  },
-  [theme.breakpoints.down('md')]: {
-    paddingRight: theme.spacing(10),
-    paddingLeft: theme.spacing(10),
-  },
-  [theme.breakpoints.down('sm')]: {
-    paddingRight: withSlider ? 0 : theme.spacing(4),
-    paddingLeft: theme.spacing(4),
-    paddingTop: isPaddingTopMobile ? theme.spacing(17.75) : 0,
-  },
-  [theme.breakpoints.up('xl')]: {
-    paddingRight: theme.spacing(5),
-    paddingLeft: theme.spacing(5),
-  },
-}));
+  withSlider?: boolean;
+  visibleOverflow?: boolean;
+  isPaddingTopMobile?: boolean;
+}>(
+  ({
+    theme,
+    withSlider = false,
+    visibleOverflow = false,
+    isPaddingTopMobile = true,
+  }) => ({
+    maxWidth: theme.breakpoints.values.lg,
+    margin: '0 auto',
+    [theme.breakpoints.up('lg')]: {
+      overflow: visibleOverflow ? 'visible' : 'hidden',
+    },
+    [theme.breakpoints.down('lg')]: {
+      overflow: 'hidden',
+    },
+    [theme.breakpoints.up('sm')]: {
+      paddingTop: theme.spacing(22.25),
+    },
+    [theme.breakpoints.up('md')]: {
+      paddingRight: theme.spacing(37.5),
+      paddingLeft: theme.spacing(37.5),
+    },
+    [theme.breakpoints.down('md')]: {
+      paddingRight: theme.spacing(10),
+      paddingLeft: theme.spacing(10),
+    },
+    [theme.breakpoints.down('sm')]: {
+      paddingRight: withSlider ? 0 : theme.spacing(4),
+      paddingLeft: theme.spacing(4),
+      paddingTop: isPaddingTopMobile ? theme.spacing(17.75) : 0,
+    },
+    [theme.breakpoints.up('xl')]: {
+      paddingRight: theme.spacing(5),
+      paddingLeft: theme.spacing(5),
+    },
+  }),
+);
 
 const useStyles = makeStyles<StyleProps>()(
   (
