--- web-marketplace/src/components/organisms/GettingStartedResourcesSection/GettingStartedResourcesSection.styles.ts
@@ -1,14 +1,0 @@
-import { makeStyles } from 'tss-react/mui';
-
-export const useSectionStyles = makeStyles()(theme => ({
-  root: {
-    [theme.breakpoints.down('sm')]: {
-      paddingTop: theme.spacing(3.5),
-    },
-  },
-  title: {
-    [theme.breakpoints.down('sm')]: {
-      textAlign: 'left',
-    },
-  },
-}));
